Safari-609.1.11.1/LayoutTests/fast/dom/DOMURL/searchparams-expected.txt

CONSOLE MESSAGE: line 8: PASS
CONSOLE MESSAGE: line 11: PASS
CONSOLE MESSAGE: line 15: PASS
CONSOLE MESSAGE: line 19: PASS
